+++ back/src/main/java/com/example/_thecore_back/emulator/infrastructure/EmulatorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EmulatorRepository extends JpaRepository<EmulatorEntity, Long> {
+public interface EmulatorRepository extends JpaRepository<EmulatorEntity, Integer> {
     // TODO: car_number -> device_id 임시 변경. 추후 car테이블 머지 후 수정 필요
     Optional<EmulatorEntity> findByDeviceId(String deviceId);
